rename(console): 패키지 명을 utils에서 adapter로 변경
